feat(sanity): add useVersionOperations, replace createVersion and discardVersion across components
refactor(sanity): add useVersionOperations, replace createVersion on dialog and chip

refactor(sanity): add discard version to useVersionOperations, replace discard version on dialog in document header

refactor(sanity): Move DiscardVersionDialog to Releases directory, export and re-use it DocumentActions
