+++ packages/sanity/src/core/releases/hooks/__tests__/useVersionOperations.test.tsx
@@ -1,0 +1,13 @@
+import {beforeEach, describe, it, vi} from 'vitest'
+
+describe('useVersionOperations', () => {
+  beforeEach(() => {
+    vi.clearAllMocks()
+  })
+
+  it.todo('should create a version successfully', async () => {})
+
+  it.todo('should discard a version successfully')
+
+  it.todo('should handle discard version failure')
+})
