+++ packages/sanity/src/core/releases/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './useDocumentVersions'
 export * from './usePerspective'
+export * from './useVersionOperations'
